com.unity.dots.runtime@0.23.0-preview ## [0.23.0] - 2020-03-03 * Added window to select what configs to generate the dots solution with. * Stopped requiring a reference to Unity.Tiny.Main and Unity.Tiny.EntryPoint for a root assembly to work. Implicitly adds those if an entry point is not referenced.
## [0.22.0] - 2020-02-05
* Add new required Dots runtime project settings build component to author the DisplayInfo runtime component (resolution, color space ..)

+++ LowLevelSupport~/Unity.LowLevel/UnsafeUtility.cs
@@ -107,7 +107,7 @@ namespace Unity.Collections.LowLevel.Unsafe
             return PinSystemArrayAndGetAddress(target, out gcHandle);
         }
 
-        public static unsafe int MemoryCompare(void* left, void* right, int size)
+        public static unsafe int MemoryCompare(void* left, void* right, long size)
         {
             byte* bl = (byte*) left;
             byte* br = (byte*) right;
